Extract out showcase halls into their own files
