+++ server/src/rooms/eastShowcaseHall.ts
@@ -1,0 +1,84 @@
+export default {
+    id: 'eastShowcaseHall',
+        displayName: 'East Showcase Hall',
+            shortName: 'the east showcase hall',
+                description: `
+      A number of booths line the sides of the hall.<br/><br/>
+      Here you can see Roundguard by Andrea Roberts, Nongunz: Doppelganger Edition by Mike Corrigan, Ultimate ADOM: Caverns of Chaos by Thomas Biskup and Team ADOM, and Cantrip by Joeri Bakker.<br/><br/>
+
+      <div id="east-showcase" class="showcase-container">
+        <div id="roundguard-entry" class="showcase-entry">
+          <p><a href="https://store.steampowered.com/app/848030/Roundguard" target="_blank">Roundguard</a>, by 
+            <a href="https://twitter.com/TanookiFruit" target="_blank" rel="nofollow noopener noreferrer">Andrea Roberts</a>
+          </p>
+          <p>Roundguard is a bouncy dungeon crawler with pinball physics, lots of loot, and a randomized castle full of 
+            oddballs. Press your luck against hordes of dangerously cute monsters and challenging roguelike elements in this 
+            all-round bouncy adventure! You can find it on 
+            <a href="https://store.steampowered.com/app/848030/Roundguard" target="_blank">Steam</a>, 
+            <a href="https://apps.apple.com/us/app/roundguard/id1484566760" target="_blank">Apple Arcade</a>, 
+            <a href="https://www.nintendo.com/games/detail/roundguard-switch/" target="_blank">Switch</a>, 
+            <a href="https://www.microsoft.com/en-us/p/roundguard/9pb3fdwrwb9l?activetab=pivot:overviewtab" target="_blank">Xbox One</a>, and 
+            <a href="https://store.playstation.com/en-us/product/UP2199-CUSA17809_00-ROUNDGUARDNA0002" target="_blank">PS4</a>!
+          </p>
+          <img src="images/roundguard/Roundguard_1.png" alt="Roundguard 1" style="width: 32%;">
+          <img src="images/roundguard/Roundguard_2.png" alt="Roundguard 2" style="width: 32%;">
+          <img src="images/roundguard/Roundguard_3.png" alt="Roundguard 3" style="width: 32%;">
+        </div>
+
+        <div id="nongunz-entry" class="showcase-entry">
+          <p><a href="https://store.steampowered.com/app/1304780/Nongunz_Doppelganger_Edition/" target="_blank">Nongunz: Doppelganger Edition</a>, by 
+            <a href="https://twitter.com/digeratiDM" target="_blank" rel="nofollow noopener noreferrer">Mike Corrigan</a>
+          </p>
+          <p>Nongunz is a roguelite action-platformer described as “The Binding of Isaac meets Cookie Clicker.” Battle grotesque 
+            creatures, kill quickly to maintain your combo meter, and rescue worshipers to become part of your idle army to buy 
+            upgrades. You can find the demo on 
+            <a href="https://store.steampowered.com/app/1304780/Nongunz_Doppelganger_Edition/" target="_blank">Steam</a>!
+          </p>
+          <img src="images/nongunz-de/NongunzDE_09_Scaled.png" alt="Nongunz DE 9" style="width: 32%;">
+          <img src="images/nongunz-de/NongunzDE_011_Scaled.png" alt="Nongunz DE 11" style="width: 32%;">
+          <img src="images/nongunz-de/NongunzDE_012_Scaled.png" alt="Nongunz DE 12" style="width: 32%;">
+        </div>
+
+        <div id="ultimate-adom-entry" class="showcase-entry">
+          <div style="display: inline-flex;height:  fit-content;">
+            <p style="width: fit-content;font-size: x-large;height: fit-content; margin: auto;">
+              <a href="https://www.ultimate-adom.com" target="_blank" style="color: magenta;">Ultimate ADOM: Caverns of Chaos</a>
+            </p>
+            <p style="padding-left: 1em; margin: auto;">by
+              <a href="https://twitter.com/adom_dev" target="_blank" rel="nofollow noopener noreferrer" style="color: magenta;">
+                Team ADOM</a>
+            </p>
+          </div>
+
+          <p>After over 25 years, Team ADOM brings the official successor to the award-winning original. Grab your blade, ready 
+            your spells and join one of the powerful factions vying for control or destruction of the source of ChAoS found 
+            deep within the ominous and deadly Caverns of Chaos.
+          </p>
+          <p>Wishlist us now on 
+            <a href="https://store.steampowered.com/app/1266820/Ultimate_ADOM__Caverns_of_Chaos" target="_blank">Steam</a> and 
+            prepare for Early Access!
+          </p>
+          <img src="images/ultimate-adom/left.png" alt="Ultimate ADOM left" style="width: 49%; max-height: 215px; object-fit: contain">
+          <img src="images/ultimate-adom/right.png" alt="Ultimate ADOM right" style="width: 49%; max-height: 215px; object-fit: contain">
+          <div style="margin: 0 auto; width: fit-content;">
+            <iframe src="https://store.steampowered.com/widget/1266820/" frameborder="0" width="646" height="190"></iframe>
+          </div>
+        </div>
+
+        <div id="cantrip-entry" class="showcase-entry">
+          <p><a href="https://bakkerjoeri.itch.io/cantrip" target="_blank">Cantrip</a>,
+              by <a href="https://twitter.com/bakkerjoeri" target="_blank" rel="nofollow noopener noreferrer">Joeri Bakker</a>
+          </p>
+          <p>Battle your way through 9 enemies and draft from 40+ unique cards to build up your deck with powerful synergies
+            in this card battler roguelite with a twist: Your hand is your health. Damage makes you discard cards,
+            and if you take a hit while your hand is empty, it's game over.
+          </p>
+          <iframe src="https://itch.io/embed/381590?linkback=true&amp;dark=true" width="552" height="167" frameborder="0"><a href="https://bakkerjoeri.itch.io/cantrip" target="_blank">Cantrip by Joeri</a></iframe>
+          <img src="images/cantrip/0.png" alt="The battle screen of Cantrip. The player's hand visible on the bottom. The opponent's hand is face down at the top. Both have a drawing pile and graveyard next to it." style="width: 49%;">
+          <img src="images/cantrip/1.png" alt="The drafting screen of Cantrip. Three cards are lined up in the middle, waiting to be picked." style="width: 49%;">
+      </div>
+      </div>
+
+      You can exit to the [[north hall->northShowcaseHall]] or [[south hall->southShowcaseHall]]`,
+                    allowsMedia: true
+},
