+++ server/src/rooms/northShowcaseHall.ts
@@ -1,0 +1,79 @@
+export default {
+    id: 'northShowcaseHall',
+        displayName: 'North Showcase Hall',
+            shortName: 'the north showcase hall',
+                description: `
+      A number of booths line the sides of the hall.<br/><br/>
+      Here you can see Gesuido by Ryosuke Mihara, Project Earth by Michael Taylor, and WizardWarsIO by quantompotato.<br/><br/>
+
+      <div id="north-showcase" class="showcase-container">
+        <div id="gesuido-entry" class="showcase-entry" style="display: inline-flex;">
+          <div id="gesuido-text">
+            <p><a href="https://mihararyosuke.com/gesuido" target="_blank">Gesuido</a>, by 
+              <a href="https://twitter.com/oinariman" target="_blank" rel="nofollow noopener noreferrer">Ryosuke Mihara</a>
+            </p>
+            <p>Gesuido is a Roguelike game for iPhone that's not too dissimilar in gameplay to the original Rogue, but it features early 
+              Macintosh-style black and white tile graphics and cute illustrated magic cards. The game is still developing, but I’ve made a beta 
+              version available for Roguelike Celebration 2020 attendees. Feedbacks are welcome. You can find it 
+              <a href="https://www.google.com" target="_blank">[beta link will be added when available]</a>!
+            </p>
+          </div>
+          <img src="images/gesuido/gesuido_ss01.png" alt="Gesuido screenshot 1" style="width: 30%; object-fit: scale-down;">
+          <img src="images/gesuido/gesuido_ss02.png" alt="Gesuido screenshot 2" style="width: 30%; object-fit: scale-down;">
+        </div>
+
+        <div id="project-earth-entry" class="showcase-entry">
+          <p><a href="https://axeforge.itch.io/project-earth" target="_blank">Project Earth</a>, by Michael Taylor</p>
+          <p>Project Earth is tactical Roguelike where instead of finding your gear, you craft your gear! Instead enemies drop 
+            materials when defeated which you use as materials for crafting. The type of materials used affects the kinds of 
+            items you make. This opens up many different tactical options. Like all Roguelikes, it's meant to be brutal and you 
+            will die a bunch of times. But when you die, what happens to the items you made? After all, they are still usable in 
+            anyone's possession...
+          </p>
+          <p>Project Earth is currently in the early stages of development. This is a playable prototype for Windows (64-bit). 
+            Currently, I'm looking for feedback on the following:
+          </p>
+          <ul>
+            <li>Combat</li>
+            <li>Crafting</li>
+            <li>World progression</li>
+          </ul>
+          <p>Feel free to mention anything else as well! Thanks for trying out Project Earth!</p>
+          <p>You can find the playable prototype <a href="https://axeforge.itch.io/project-earth" target="_blank">here</a> on 
+            itch.io. Instructions on how to play are on the site.
+          </p>
+          <p>I can be contacted in the following ways:</p>
+          <ul>
+            <li>Discord: AxeForge #0707 I'm on the following discord servers as well:
+              <ul>
+                <li>roguelikes</li>
+                <li>Tangledeep</li>
+                <li>Dungeonfans Academy</li>
+                <li>Auto Combat Zone</li>
+              </ul>
+            </li>
+            <li>Leaving a comment on Project Earth's itch.io page.</li>
+          </ul>
+        </div>
+
+        <div id="wizard-wars-io-entry" class="showcase-entry">
+          <h3><a href="https://github.com/WizardWarsIO" target="_blank">WizardWars.IO</a>, by R. Riley Holmes & quantumpotato</h3>
+          <strong>Rhythmic turnbased deathmatch!</strong>
+          <div>
+            <p>Multiplayer roguelike combat. Toss potions, zap wands, detonate bombs, dodge minotaurs and kill players to crown 
+              the leaderboard. <a href="http://206.189.227.251:8000" target="_blank">Play it online!</a>
+            </p>
+            <p>Press a number 1-9 to select an item. Press F + Arrow Keys to zap a wand or throw a potion. Press A to equip 
+              selected item. You can move and throw in 8 directions.
+            </p>
+          </div>
+          <img src="images/wizard-wars-io/0.png" alt="wizard wars io 0" style="width: 23%; object-fit: scale-down; max-height: 340px">
+          <img src="images/wizard-wars-io/1.png" alt="wizard wars io 1" style="width: 23%; object-fit: scale-down; max-height: 340px">
+          <img src="images/wizard-wars-io/2.png" alt="wizard wars io 2" style="width: 23%; object-fit: scale-down; max-height: 340px">
+          <img src="images/wizard-wars-io/3.png" alt="wizard wars io 3" style="width: 23%; object-fit: scale-down; max-height: 340px">
+        </div>
+      </div>
+
+
+      You can exit to the [[west hall->westShowcaseHall]] or [[east hall->eastShowcaseHall]]`,
+                    allowsMedia: true
