+++ server/src/rooms/southShowcaseHall.ts
@@ -1,0 +1,77 @@
+export default {
+    id: 'southShowcaseHall',
+    displayName: 'South Showcase Hall',
+    shortName: 'the south showcase hall',
+    description: `
+      A number of booths line the sides of the hall.<br/><br/>
+      Here you can see Fuzz Force: Spook Squad by Alan Igle, Mech@mor Showdown by Seth Alter, Rift Wizard by Dylan White, and Happy Grumps by Glenn LaBarre.<br/><br/>
+
+      <div id="south-showcase" class="showcase-container">
+        <div id="fuzz-force-entry" class="showcase-entry">
+          <p><a href="https://fuzzforce.itch.io/fuzz-force-spook-squad" target="_blank">Fuzz Force: Spook Squad</a>, by 
+            <a href="https://twitter.com/FuzzForce" target="_blank" rel="nofollow noopener noreferrer">Alan Igle</a>
+          </p>
+          <p> Deckbuilding meets Dice in a quest to capture the ghoulish Polter Prince. Join the Spook Squad and upgrade their 
+            Dice-powered weapons to give those ghosts something to be scared of! You can find a demo on 
+            <a href="https://fuzzforce.itch.io/fuzz-force-spook-squad" target="_blank">itch.io</a>!
+          </p>
+          <img src="images/fuzz-force/Finn_Dice_Web.jpg" alt="Finn_Dice_Web" style="width: 49%;">
+          <img src="images/fuzz-force/ForestLevel_Web.jpg" alt="ForestLevel_Web" style="width: 49%;">
+          <img src="images/fuzz-force/DiceSwap_Web.jpg" alt="DiceSwap_Web" style="width: 49%;">
+          <img src="images/fuzz-force/FerretyShop_Dot_Web.jpg" alt="ForretyShop_Dot_Web" style="width: 49%;">
+        </div>
+
+        <div id="mech@mor-showdown-entry" class="showcase-entry">
+          <p><a href="https://subalterngames.itch.io/mechamor-showdown" target="_blank">Mech@mor Showdown</a>, by 
+            <a href="https://twitter.com/subalterngames" target="_blank" rel="nofollow noopener noreferrer">Seth Alter</a>
+          </p>
+          <p>Mech@mor showdown is an ASCII roguelike about your relationships with your mech and with your opponents. Build 
+            trust between you and your mecha. Fight your foes. Flirt with your rivals. Taunt your lovers. Swap out parts for 
+            better parts. Save Planet Yendor for the regime of the Kyriarch.
+          </p>
+          <p> Pre-release available at <a href="https://subalterngames.itch.io/mechamor-showdown" target="_blank">itch.io</a>!</p>
+          <img src="images/mech@mor-showdown/0.png" alt="Mech@amor Showdown 0" style="width: 49%;">
+          <img src="images/mech@mor-showdown/1.png" alt="Mech@amor Showdown 1" style="width: 49%;">
+          <img src="images/mech@mor-showdown/2.png" alt="Mech@amor Showdown 2" style="width: 49%;">
+          <img src="images/mech@mor-showdown/3.png" alt="Mech@amor Showdown 3" style="width: 49%;">
+        </div>
+
+        <div id="rift-wizard-entry" class="showcase-entry">
+          <p><a href="https://store.steampowered.com/app/1271280/Rift_Wizard" target="_blank">Rift Wizard</a>, by
+            <a href="https://twitter.com/GameDevDylanW" target="_blank" rel="nofollow noopener noreferrer">Dylan White</a>
+          </p>
+          <p>Rift Wizard is a tactical spellcasting roguelike. The game emphasizes combat and character builds. The character system allows the 
+            player extreme freedom, with many possible synergies and combos to be discovered and exploited. Levels are compact, fitting entirely 
+            on one screen, allowing the player to fully plan out a clever strategy for each battle. Diverse enemies, high impact random loot, and 
+            procedurally generated battlefields ensure that each run is a unique strategic challenge.
+          </p>
+          <img src="images/rift-wizard/ss_16.png" alt="Rift Wizard screenshot 16" style="width: 49%; height: 150px; object-fit: none; object-position:center;">
+          <img src="images/rift-wizard/ss_17.png" alt="Rift Wizard screenshot 17" style="width: 49%; height: 150px; object-fit: none; object-position:center;">
+          <div style="margin: 0 auto; width: fit-content;">
+            <iframe src="https://store.steampowered.com/widget/1271280/" frameborder="0" width="646" height="190"></iframe>
+          </div>
+        </div>
+
+        <div id="happy-grumps-entry" class="showcase-entry">
+          <div style="margin: auto; width: 50%; height: 200px;">
+            <div style="display: inline-flex; margin: auto;">
+              <a href="https://store.steampowered.com/app/1109170/Happy_Grumps rel="nofollow noopener noreferrer"" target="_blank">
+                <img src="images/happy-grumps/logo.png" alt="Happy Grumps Logo" style="max-height: 200px;">
+              </a>
+              <p style="margin: auto;">by 
+                <a href="https://twitter.com/HappyGrumps" target="_blank" rel="nofollow noopener noreferrer">Glenn LaBarre</a>
+              </p>
+            </div>
+          </div>
+          <p>Happy Grumps is a love letter to classic roguelike adventures with a twist. Instead of trying to steal treasure or 
+            fight monsters, you are on a quest to spread happiness to a grumpy world. You can find it on 
+            <a href="https://store.steampowered.com/app/1109170/Happy_Grumps rel="nofollow noopener noreferrer"" target="_blank">Steam</a>!
+          </p>
+          <img src="images/happy-grumps/0.png" alt="Happy Grumps 0" style="width: 49%;">
+          <img src="images/happy-grumps/1.png" alt="Happy Grumps 1" style="width: 49%;">
+        </div>
+      </div>
+
+      You can exit to the [[west hall->westShowcaseHall]] or [[east hall->eastShowcaseHall]]`,
+    allowsMedia: true
+},
