+++ server/src/rooms/westShowcaseHall.ts
@@ -1,0 +1,63 @@
+export default {
+  id: 'westShowcaseHall',
+  displayName: 'West Showcase Hall',
+  shortName: 'the west showcase hall',
+  description: `
+      A number of booths line the sides of the hall.<br/><br/>
+      Here you can see Computer Dungeon Slash: ZZT by KKairos, Peglin by Red Nexus Games, and AutoFire by Patrick Lipo.<br/><br/>
+
+      <div class="showcase-container">
+        <div id="cdslash-entry" class="showcase-entry">
+          <p><a href="https://kkairos.itch.io/cdslash" target="_blank">Computer Dungeon Slash: ZZT</a>, by 
+            <a href="https://www.kaikairos.dev" target="_blank" rel="nofollow noopener noreferrer">KKairos</a>
+          </p>
+          <p>Computer Dungeon Slash: ZZT is a "dungeon crawler" written for the 1991 game-creation system ZZT featuring classic "ZZT-style" action, 
+            dynamic procedural level generation, and a comedic cast of characters to rescue. You can find it on 
+            <a href="https://kkairos.itch.io/cdslash" target="_blank">itch.io</a>!
+          </p>
+          <img src="images/cdszzt/cdszzt-title.png" alt="The Computer Dungeon Slash: ZZT title screen" style="max-width: 32%;">
+          <img src="images/cdszzt/cdszzt-town.png" alt="The town area of Computer Dungeon Slash: ZZT" style="max-width: 32%;">
+          <img src="images/cdszzt/cdszzt-montage.png" alt="Several different 'limited viewport' levels of Computer Dungeon Slash: ZZT" style="max-width: 32%;">
+        </div>
+
+        <div id="peglin-entry" class="showcase-entry">
+          <p><a href="https://store.steampowered.com/app/1296610/Peglin" target="_blank">Peglin</a>, by 
+            <a href="https://twitter.com/rednexusgames" target="_blank" rel="nofollow noopener noreferrer">Red Nexus Games</a>
+          </p>
+          <p>Peglin is a Pachinko Roguelike: Fight enemies by collecting special orbs and popping pegs to deal damage. Acquire 
+            special relics that radically change the stakes and ensure no two runs are the same. Aim carefully to survive in 
+            this unique turn-based RPG! You can find it on 
+            <a href="https://store.steampowered.com/app/1296610/Peglin" target="_blank">Steam</a>!
+          </p>
+          <div style="width: 100%">
+            <video class="showcase-video" controls style="margin: auto; width: 420px; height: 240px;">
+                <source src="images/peglin/peglin_video.mp4" type="video/mp4">
+                <source src="movie.ogg" type="video/ogg">
+              Your browser does not support the video tag.
+            </video>
+          </div>
+          <img src="images/peglin/peglin1.png" alt="Peglin 1" style="max-width: 32%;">
+          <img src="images/peglin/peglin2.png" alt="Peglin 2" style="max-width: 32%;">
+          <img src="images/peglin/peglin3.png" alt="Peglin 3" style="max-width: 32%;">
+        </div>
+
+        <div id="auto-fire-entry" class="showcase-entry">
+          <p><a href="https://vertigames.itch.io/auto-fire" target="_blank">Auto Fire</a>, by 
+            <a href="https://twitter.com/autofiregame" target="_blank" rel="nofollow noopener noreferrer">Patrick Lipo</a>
+          </p>
+          <p><strong>Auto Fire</strong> combines classic car combat with a Tactical RPG that roams across outlaw lands in a world 
+            where drivers are worshiped like celebrities. Salvage equipment and cars and build the sexiest, deadliest ride. 
+            Work for a corp, haul cargo, fight in the combat arena or become a savior. The choice is yours. You can play the 
+            demo at <a href="https://vertigames.itch.io/auto-fire" target="_blank">itch.io</a>, and wishlist it on 
+            <a href="https://store.steampowered.com/app/1147880/Auto_Fire/">Steam</a>!
+          </p>
+          <img src="images/auto-fire/AutoFireDesertTarget1280.jpg" alt="Auto Fire Desert Target" style="max-width: 49%;">
+          <img src="images/auto-fire/AutoFireGarage1280.jpg" alt="Auto Fire Garage" style="max-width: 49%;">
+          <img src="images/auto-fire/AutoFireOverworld1280.jpg" alt="Auto Fire Overworld" style="max-width: 49%;">
+          <img src="images/auto-fire/AutoFireRuinsCars1280.jpg" alt="Auto Fire RuinsCars" style="max-width: 49%;">
+        </div>
+      </div>
+
+      You can exit to the [[north hall->northShowcaseHall]] or [[south hall->southShowcaseHall]]`,
+  allowsMedia: true
+}
