+++ server/src/types.ts
@@ -1,4 +1,4 @@
-import { Room } from './room'
+import { Room } from './rooms'
 import { User, MinimalUser, PublicUser } from './user'
 import { RoomNote } from './roomNote'
